Fixed linter issue with circular dependencies
