+++ packages/framework/src/state/screen.ts
@@ -1,17 +1,7 @@
-import { atom, useAtom, useAtomValue } from "jotai";
+import { atom } from "jotai";
 import { focusAtom } from "jotai-optics";
-import { clone } from "lodash-es";
-import { useCallback } from "react";
-import isEqual from "react-fast-compare";
 import type { Response, WebSocketConnection } from "../data";
-import type {
-  EnsembleAPIModel,
-  EnsembleAppModel,
-  EnsembleMockResponse,
-  EnsembleScreenModel,
-  EnsembleSocketModel,
-} from "../shared";
-import { useEvaluate } from "../hooks";
+import type { EnsembleAppModel, EnsembleScreenModel } from "../shared";
 import type { WidgetState } from "./widget";
 
 export interface ScreenContextDefinition {
@@ -62,48 +52,3 @@ export const screenGlobalScriptAtom = focusAtom(screenAtom, (optic) =>
 export const screenImportScriptAtom = focusAtom(screenAtom, (optic) =>
   optic.prop("model").optional().prop("importedScripts"),
 );
-
-export const useScreenData = (): { apis?: EnsembleAPIModel[] } & {
-  sockets?: EnsembleSocketModel[];
-} & Pick<ScreenContextDefinition, "data"> &
-  Pick<ScreenContextActions, "setData"> & {
-    mockResponses: { [key: string]: EnsembleMockResponse | string | undefined };
-  } => {
-  const apis = useAtomValue(screenApiAtom);
-  const sockets = useAtomValue(screenSocketAtom);
-  const [data, setDataAtom] = useAtom(screenDataAtom);
-
-  const mockResponses = useEvaluate(
-    apis
-      ? apis.reduce(
-          (
-            acc: { [key: string]: EnsembleMockResponse | string | undefined },
-            api,
-          ) => {
-            acc[api.name] = api.mockResponse;
-            return acc;
-          },
-          {},
-        )
-      : {},
-  );
-
-  const setData = useCallback(
-    (name: string, response: Response | WebSocketConnection) => {
-      if (isEqual(data[name], response)) {
-        return;
-      }
-      data[name] = response;
-      setDataAtom(clone(data));
-    },
-    [data, setDataAtom],
-  );
-
-  return {
-    apis,
-    sockets,
-    data,
-    setData,
-    mockResponses,
-  };
-};
